add ability to use but not commit service key data
We want to use service key data in JSON files, but not have them committed to the repository. So we have a new convention where service key JSON files are not <instance>-<servicekey>.json but sk-<instance>-<servicekey>.json so we can pick them out and use the pattern in .gitignore.